feat: add realtime recon
